+++ test/e2e/io_soak/README.md
@@ -12,3 +12,16 @@ Runs fio with varying duty cycles concurrently on a number of volumes for an ext
 `e2e_io_soak_duration` is parsed using `golangs` library function `time.ParseDuration`.
 So `e2e_io_soak_duration` string is a sequence of decimal numbers, each with optional fraction and a unit suffix, such as "300ms", "-1.5h" or "2h45m".
 Valid time units are "ns", "us" (or "µs"), "ms", "s", "m", "h". 
+
+## Notes:
+To facilitate creation of a largish number of test pods in a reasonable amount of time,
+all test pods (excluding the disruptor pods) run fio with a delay of 60 seconds,
+this is an attempt to makes the test setup more robust.
+The delay means CPU utilization on the cluster nodes for the initial 60 seconds is low
+and should make it possible to create pods at a reasonable rate.
+
+All fio runs including the disruptor pods run fio with verification on,
+though the disruptor fio run is not checked.
+
+The disruptor pods are identical to test pods execpt that they been configured
+to raise a SIGSEGV after a configurable delay
